% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Acl}
\alias{Acl}
\title{Google+ API Objects 
Builds on top of the Google+ platform.}
\usage{
Acl(description = NULL, items = NULL)
}
\arguments{
\item{description}{Description of the access granted, suitable for display}

\item{items}{The list of access entries}
}
\value{
Acl object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:04:45
filename: /Users/mark/dev/R/autoGoogleAPI/googleplusv1.auto/R/plus_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Acl Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
