% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_functions.R
\name{activities.get}
\alias{activities.get}
\title{Get an activity.}
\usage{
activities.get(activityId)
}
\arguments{
\item{activityId}{The ID of the activity to get}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/plus.me
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/plus.login, https://www.googleapis.com/auth/plus.me)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/+/api/}{Google Documentation}
}
