% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Activity.actor.clientSpecificActorInfo}
\alias{Activity.actor.clientSpecificActorInfo}
\title{Activity.actor.clientSpecificActorInfo Object}
\usage{

  Activity.actor.clientSpecificActorInfo(Activity.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  youtubeActorInfo = NULL)
}
\arguments{
\item{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{youtubeActorInfo}{Actor info specific to YouTube clients}
}
\value{
Activity.actor.clientSpecificActorInfo object
}
\description{
Activity.actor.clientSpecificActorInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Actor info specific to particular clients.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}}
}

