% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Activity.object.attachments.thumbnails.image}
\alias{Activity.object.attachments.thumbnails.image}
\title{Activity.object.attachments.thumbnails.image Object}
\usage{
Activity.object.attachments.thumbnails.image(height = NULL, type = NULL,
  url = NULL, width = NULL)
}
\arguments{
\item{height}{The height, in pixels, of the linked resource}

\item{type}{Media type of the link}

\item{url}{Image url}

\item{width}{The width, in pixels, of the linked resource}
}
\value{
Activity.object.attachments.thumbnails.image object
}
\description{
Activity.object.attachments.thumbnails.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Image resource.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}}
}
