% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Activity.object.replies}
\alias{Activity.object.replies}
\title{Activity.object.replies Object}
\usage{
Activity.object.replies(selfLink = NULL, totalItems = NULL)
}
\arguments{
\item{selfLink}{The URL for the collection of comments in reply to this activity}

\item{totalItems}{Total number of comments on this activity}
}
\value{
Activity.object.replies object
}
\description{
Activity.object.replies Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Comments in reply to this activity.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}}
}

