% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Comment.actor.clientSpecificActorInfo}
\alias{Comment.actor.clientSpecificActorInfo}
\title{Comment.actor.clientSpecificActorInfo Object}
\usage{

  Comment.actor.clientSpecificActorInfo(Comment.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  youtubeActorInfo = NULL)
}
\arguments{
\item{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{youtubeActorInfo}{Actor info specific to YouTube clients}
}
\value{
Comment.actor.clientSpecificActorInfo object
}
\description{
Comment.actor.clientSpecificActorInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Actor info specific to particular clients.
}
\seealso{
Other Comment functions: \code{\link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Comment.actor.image}},
  \code{\link{Comment.actor.verification}},
  \code{\link{Comment.actor}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.object}},
  \code{\link{Comment.plusoners}}, \code{\link{Comment}}
}

