% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}
\alias{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}
\title{Comment.actor.clientSpecificActorInfo.youtubeActorInfo Object}
\usage{
Comment.actor.clientSpecificActorInfo.youtubeActorInfo(channelId = NULL)
}
\arguments{
\item{channelId}{ID of the YouTube channel owned by the Actor}
}
\value{
Comment.actor.clientSpecificActorInfo.youtubeActorInfo object
}
\description{
Comment.actor.clientSpecificActorInfo.youtubeActorInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Actor info specific to YouTube clients.
}
\seealso{
Other Comment functions: \code{\link{Comment.actor.clientSpecificActorInfo}},
  \code{\link{Comment.actor.image}},
  \code{\link{Comment.actor.verification}},
  \code{\link{Comment.actor}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.object}},
  \code{\link{Comment.plusoners}}, \code{\link{Comment}}
}
