% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Comment.actor}
\alias{Comment.actor}
\title{Comment.actor Object}
\usage{
Comment.actor(Comment.actor.clientSpecificActorInfo = NULL,
  Comment.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  Comment.actor.image = NULL, Comment.actor.verification = NULL,
  clientSpecificActorInfo = NULL, displayName = NULL, id = NULL,
  image = NULL, url = NULL, verification = NULL)
}
\arguments{
\item{Comment.actor.clientSpecificActorInfo}{The \link{Comment.actor.clientSpecificActorInfo} object or list of objects}

\item{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{Comment.actor.image}{The \link{Comment.actor.image} object or list of objects}

\item{Comment.actor.verification}{The \link{Comment.actor.verification} object or list of objects}

\item{clientSpecificActorInfo}{Actor info specific to particular clients}

\item{displayName}{The name of this actor, suitable for display}

\item{id}{The ID of the actor}

\item{image}{The image representation of this actor}

\item{url}{A link to the Person resource for this actor}

\item{verification}{Verification status of actor}
}
\value{
Comment.actor object
}
\description{
Comment.actor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The person who posted this comment.
}
\seealso{
Other Comment functions: \code{\link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Comment.actor.clientSpecificActorInfo}},
  \code{\link{Comment.actor.image}},
  \code{\link{Comment.actor.verification}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.object}},
  \code{\link{Comment.plusoners}}, \code{\link{Comment}}
}
