% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Comment.actor.verification}
\alias{Comment.actor.verification}
\title{Comment.actor.verification Object}
\usage{
Comment.actor.verification(adHocVerified = NULL)
}
\arguments{
\item{adHocVerified}{Verification for one-time or manual processes}
}
\value{
Comment.actor.verification object
}
\description{
Comment.actor.verification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Verification status of actor.
}
\seealso{
Other Comment functions: \code{\link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Comment.actor.clientSpecificActorInfo}},
  \code{\link{Comment.actor.image}},
  \code{\link{Comment.actor}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.object}},
  \code{\link{Comment.plusoners}}, \code{\link{Comment}}
}
