% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Comment}
\alias{Comment}
\title{Comment Object}
\usage{
Comment(Comment.actor = NULL, Comment.actor.clientSpecificActorInfo = NULL,
  Comment.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  Comment.actor.image = NULL, Comment.actor.verification = NULL,
  Comment.inReplyTo = NULL, Comment.object = NULL,
  Comment.plusoners = NULL, actor = NULL, etag = NULL, id = NULL,
  inReplyTo = NULL, object = NULL, plusoners = NULL, published = NULL,
  selfLink = NULL, updated = NULL, verb = NULL)
}
\arguments{
\item{Comment.actor}{The \link{Comment.actor} object or list of objects}

\item{Comment.actor.clientSpecificActorInfo}{The \link{Comment.actor.clientSpecificActorInfo} object or list of objects}

\item{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{Comment.actor.image}{The \link{Comment.actor.image} object or list of objects}

\item{Comment.actor.verification}{The \link{Comment.actor.verification} object or list of objects}

\item{Comment.inReplyTo}{The \link{Comment.inReplyTo} object or list of objects}

\item{Comment.object}{The \link{Comment.object} object or list of objects}

\item{Comment.plusoners}{The \link{Comment.plusoners} object or list of objects}

\item{actor}{The person who posted this comment}

\item{etag}{ETag of this response for caching purposes}

\item{id}{The ID of this comment}

\item{inReplyTo}{The activity this comment replied to}

\item{object}{The object of this comment}

\item{plusoners}{People who +1'd this comment}

\item{published}{The time at which this comment was initially published}

\item{selfLink}{Link to this comment resource}

\item{updated}{The time at which this comment was last updated}

\item{verb}{This comment's verb, indicating what action was performed}
}
\value{
Comment object
}
\description{
Comment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Comment functions: \code{\link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Comment.actor.clientSpecificActorInfo}},
  \code{\link{Comment.actor.image}},
  \code{\link{Comment.actor.verification}},
  \code{\link{Comment.actor}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.object}},
  \code{\link{Comment.plusoners}}
}

