% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_functions.R
\name{people.search}
\alias{people.search}
\title{Search all public profiles.}
\usage{
people.search(query, language = NULL, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{query}{Specify a query string for full text search of public text in all profiles}

\item{language}{Specify the preferred language to search with}

\item{maxResults}{The maximum number of people to include in the response, which is used for paging}

\item{pageToken}{The continuation token, which is used to page through large result sets}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/plus.me
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/plus.login, https://www.googleapis.com/auth/plus.me)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/+/api/}{Google Documentation}
}

