% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person.ageRange}
\alias{Person.ageRange}
\title{Person.ageRange Object}
\usage{
Person.ageRange(max = NULL, min = NULL)
}
\arguments{
\item{max}{The age range's upper bound, if any}

\item{min}{The age range's lower bound, if any}
}
\value{
Person.ageRange object
}
\description{
Person.ageRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The age range of the person. Valid ranges are 17 or younger, 18 to 20, and 21 or older. Age is determined from the user's birthday using Western age reckoning.
}
\seealso{
Other Person functions: \code{\link{Person.cover.coverInfo}},
  \code{\link{Person.cover.coverPhoto}},
  \code{\link{Person.cover}}, \code{\link{Person.emails}},
  \code{\link{Person.image}}, \code{\link{Person.name}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}},
  \code{\link{Person.urls}}, \code{\link{Person}}
}

