% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person.cover.coverInfo}
\alias{Person.cover.coverInfo}
\title{Person.cover.coverInfo Object}
\usage{
Person.cover.coverInfo(leftImageOffset = NULL, topImageOffset = NULL)
}
\arguments{
\item{leftImageOffset}{The difference between the left position of the cover image and the actual displayed cover image}

\item{topImageOffset}{The difference between the top position of the cover image and the actual displayed cover image}
}
\value{
Person.cover.coverInfo object
}
\description{
Person.cover.coverInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Extra information about the cover photo.
}
\seealso{
Other Person functions: \code{\link{Person.ageRange}},
  \code{\link{Person.cover.coverPhoto}},
  \code{\link{Person.cover}}, \code{\link{Person.emails}},
  \code{\link{Person.image}}, \code{\link{Person.name}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}},
  \code{\link{Person.urls}}, \code{\link{Person}}
}

