% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person.cover.coverPhoto}
\alias{Person.cover.coverPhoto}
\title{Person.cover.coverPhoto Object}
\usage{
Person.cover.coverPhoto(height = NULL, url = NULL, width = NULL)
}
\arguments{
\item{height}{The height of the image}

\item{url}{The URL of the image}

\item{width}{The width of the image}
}
\value{
Person.cover.coverPhoto object
}
\description{
Person.cover.coverPhoto Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The person's primary cover image.
}
\seealso{
Other Person functions: \code{\link{Person.ageRange}},
  \code{\link{Person.cover.coverInfo}},
  \code{\link{Person.cover}}, \code{\link{Person.emails}},
  \code{\link{Person.image}}, \code{\link{Person.name}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}},
  \code{\link{Person.urls}}, \code{\link{Person}}
}
