% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person.cover}
\alias{Person.cover}
\title{Person.cover Object}
\usage{
Person.cover(Person.cover.coverInfo = NULL, Person.cover.coverPhoto = NULL,
  coverInfo = NULL, coverPhoto = NULL, layout = NULL)
}
\arguments{
\item{Person.cover.coverInfo}{The \link{Person.cover.coverInfo} object or list of objects}

\item{Person.cover.coverPhoto}{The \link{Person.cover.coverPhoto} object or list of objects}

\item{coverInfo}{Extra information about the cover photo}

\item{coverPhoto}{The person's primary cover image}

\item{layout}{The layout of the cover art}
}
\value{
Person.cover object
}
\description{
Person.cover Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The cover photo content.
}
\seealso{
Other Person functions: \code{\link{Person.ageRange}},
  \code{\link{Person.cover.coverInfo}},
  \code{\link{Person.cover.coverPhoto}},
  \code{\link{Person.emails}}, \code{\link{Person.image}},
  \code{\link{Person.name}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}},
  \code{\link{Person.urls}}, \code{\link{Person}}
}

