% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person.image}
\alias{Person.image}
\title{Person.image Object}
\usage{
Person.image(isDefault = NULL, url = NULL)
}
\arguments{
\item{isDefault}{Whether the person's profile photo is the default one}

\item{url}{The URL of the person's profile photo}
}
\value{
Person.image object
}
\description{
Person.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The representation of the person's profile photo.
}
\seealso{
Other Person functions: \code{\link{Person.ageRange}},
  \code{\link{Person.cover.coverInfo}},
  \code{\link{Person.cover.coverPhoto}},
  \code{\link{Person.cover}}, \code{\link{Person.emails}},
  \code{\link{Person.name}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}},
  \code{\link{Person.urls}}, \code{\link{Person}}
}

