% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person.name}
\alias{Person.name}
\title{Person.name Object}
\usage{
Person.name(familyName = NULL, formatted = NULL, givenName = NULL,
  honorificPrefix = NULL, honorificSuffix = NULL, middleName = NULL)
}
\arguments{
\item{familyName}{The family name (last name) of this person}

\item{formatted}{The full name of this person, including middle names, suffixes, etc}

\item{givenName}{The given name (first name) of this person}

\item{honorificPrefix}{The honorific prefixes (such as 'Dr}

\item{honorificSuffix}{The honorific suffixes (such as 'Jr}

\item{middleName}{The middle name of this person}
}
\value{
Person.name object
}
\description{
Person.name Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object representation of the individual components of a person's name.
}
\seealso{
Other Person functions: \code{\link{Person.ageRange}},
  \code{\link{Person.cover.coverInfo}},
  \code{\link{Person.cover.coverPhoto}},
  \code{\link{Person.cover}}, \code{\link{Person.emails}},
  \code{\link{Person.image}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}},
  \code{\link{Person.urls}}, \code{\link{Person}}
}

