% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person}
\alias{Person}
\title{Person Object}
\usage{
Person(Person.ageRange = NULL, Person.cover = NULL,
  Person.cover.coverInfo = NULL, Person.cover.coverPhoto = NULL,
  Person.emails = NULL, Person.image = NULL, Person.name = NULL,
  Person.organizations = NULL, Person.placesLived = NULL,
  Person.urls = NULL, aboutMe = NULL, ageRange = NULL, birthday = NULL,
  braggingRights = NULL, circledByCount = NULL, cover = NULL,
  currentLocation = NULL, displayName = NULL, domain = NULL,
  emails = NULL, etag = NULL, gender = NULL, id = NULL, image = NULL,
  isPlusUser = NULL, language = NULL, name = NULL, nickname = NULL,
  objectType = NULL, occupation = NULL, organizations = NULL,
  placesLived = NULL, plusOneCount = NULL, relationshipStatus = NULL,
  skills = NULL, tagline = NULL, url = NULL, urls = NULL,
  verified = NULL)
}
\arguments{
\item{Person.ageRange}{The \link{Person.ageRange} object or list of objects}

\item{Person.cover}{The \link{Person.cover} object or list of objects}

\item{Person.cover.coverInfo}{The \link{Person.cover.coverInfo} object or list of objects}

\item{Person.cover.coverPhoto}{The \link{Person.cover.coverPhoto} object or list of objects}

\item{Person.emails}{The \link{Person.emails} object or list of objects}

\item{Person.image}{The \link{Person.image} object or list of objects}

\item{Person.name}{The \link{Person.name} object or list of objects}

\item{Person.organizations}{The \link{Person.organizations} object or list of objects}

\item{Person.placesLived}{The \link{Person.placesLived} object or list of objects}

\item{Person.urls}{The \link{Person.urls} object or list of objects}

\item{aboutMe}{A short biography for this person}

\item{ageRange}{The age range of the person}

\item{birthday}{The person's date of birth, represented as YYYY-MM-DD}

\item{braggingRights}{The 'bragging rights' line of this person}

\item{circledByCount}{For followers who are visible, the number of people who have added this person or page to a circle}

\item{cover}{The cover photo content}

\item{currentLocation}{(this field is not currently used)}

\item{displayName}{The name of this person, which is suitable for display}

\item{domain}{The hosted domain name for the user's Google Apps account}

\item{emails}{A list of email addresses that this person has, including their Google account email address, and the public verified email addresses on their Google+ profile}

\item{etag}{ETag of this response for caching purposes}

\item{gender}{The person's gender}

\item{id}{The ID of this person}

\item{image}{The representation of the person's profile photo}

\item{isPlusUser}{Whether this user has signed up for Google+}

\item{language}{The user's preferred language for rendering}

\item{name}{An object representation of the individual components of a person's name}

\item{nickname}{The nickname of this person}

\item{objectType}{Type of person within Google+}

\item{occupation}{The occupation of this person}

\item{organizations}{A list of current or past organizations with which this person is associated}

\item{placesLived}{A list of places where this person has lived}

\item{plusOneCount}{If a Google+ Page, the number of people who have +1'd this page}

\item{relationshipStatus}{The person's relationship status}

\item{skills}{The person's skills}

\item{tagline}{The brief description (tagline) of this person}

\item{url}{The URL of this person's profile}

\item{urls}{A list of URLs for this person}

\item{verified}{Whether the person or Google+ Page has been verified}
}
\value{
Person object
}
\description{
Person Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Person functions: \code{\link{Person.ageRange}},
  \code{\link{Person.cover.coverInfo}},
  \code{\link{Person.cover.coverPhoto}},
  \code{\link{Person.cover}}, \code{\link{Person.emails}},
  \code{\link{Person.image}}, \code{\link{Person.name}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}},
  \code{\link{Person.urls}}
}
