% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Person.urls}
\alias{Person.urls}
\title{Person.urls Object}
\usage{
Person.urls()
}
\value{
Person.urls object
}
\description{
Person.urls Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Person functions: \code{\link{Person.ageRange}},
  \code{\link{Person.cover.coverInfo}},
  \code{\link{Person.cover.coverPhoto}},
  \code{\link{Person.cover}}, \code{\link{Person.emails}},
  \code{\link{Person.image}}, \code{\link{Person.name}},
  \code{\link{Person.organizations}},
  \code{\link{Person.placesLived}}, \code{\link{Person}}
}

