% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Place.address}
\alias{Place.address}
\title{Place.address Object}
\usage{
Place.address(formatted = NULL)
}
\arguments{
\item{formatted}{The formatted address for display}
}
\value{
Place.address object
}
\description{
Place.address Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The physical address of the place.
}
\seealso{
Other Place functions: \code{\link{Place.position}},
  \code{\link{Place}}
}

