% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{Place.position}
\alias{Place.position}
\title{Place.position Object}
\usage{
Place.position(latitude = NULL, longitude = NULL)
}
\arguments{
\item{latitude}{The latitude of this position}

\item{longitude}{The longitude of this position}
}
\value{
Place.position object
}
\description{
Place.position Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The position of the place.
}
\seealso{
Other Place functions: \code{\link{Place.address}},
  \code{\link{Place}}
}
