% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_functions.R
\docType{package}
\name{plus_googleAuthR}
\alias{plus_googleAuthR}
\alias{plus_googleAuthR-package}
\title{Google+ API
Builds on top of the Google+ platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:04:45
filename: /Users/mark/dev/R/autoGoogleAPI/googleplusv1.auto/R/plus_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/plus.me
\item https://www.googleapis.com/auth/userinfo.email
\item https://www.googleapis.com/auth/userinfo.profile
}
}
