% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_objects.R
\name{PlusAclentryResource}
\alias{PlusAclentryResource}
\title{PlusAclentryResource Object}
\usage{
PlusAclentryResource(displayName = NULL, id = NULL, type = NULL)
}
\arguments{
\item{displayName}{A descriptive name for this entry}

\item{id}{The ID of the entry}

\item{type}{The type of entry describing to whom access is granted}
}
\value{
PlusAclentryResource object
}
\description{
PlusAclentryResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

